fix(nika): descend repair provenance under the crate cap
Move path classification and RunSource into nika-cli-host so nika-cli
stays under fifteen thousand production lines. Split the check, render
and run files that crossed the per-file wall, and drop expect from the
lint fixture walker.
