+++ crates/nika-cli-host/src/lib.rs
@@ -34,7 +34,9 @@ pub mod metrics;
 pub mod notify;
 pub mod output;
 pub mod probe;
+pub mod repair;
 pub mod retention;
+pub mod source;
 pub mod text;
 pub mod welcome;
 pub mod wire;
