+++ crates/nika-cli/src/registry_args.rs
@@ -1,13 +1,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 // Copyright (C) 2024-2026 SuperNovae Studio <contact@supernovae.studio>
 
-//! The `registry:` arg seam of the binary (issue #452) — swap any
-//! `registry:owner/name[@version]` file argument for its verified local
-//! cache file, then hand check/run the path exactly as if the user had
-//! typed it. A bin-side sibling of `main.rs` (the dispatcher stays a
-//! thin verb tree; this is the one arg transform it owes the registry
-//! lane). Resolution is CLI-level and happens BEFORE any workflow is
-//! parsed: a workflow's `permits:` govern its run, never this fetch.
+//! `registry:` arg seam: resolve the ref, then check/run the cached file.
 
 use nika_cli::registry;
 use nika_cli::verbs::{self, VerbOutput};
@@ -16,8 +10,7 @@ use crate::{ColorWhenArg, RunArgs, emit, run_verb};
 use nika_cli::display::format::LinkChoice;
 use nika_cli::verbs::check::{CheckFlags, CheckTarget, dispatch_targets as check_dispatch};
 
-/// The `check` arm: registry refs resolve first (the `--fix` guard
-/// rides the same seam), then the normal multi-file dispatch.
+/// `check` arm: resolve `registry:` refs, then dispatch.
 pub(crate) fn check_verb(
     files: &[String],
     flags: &CheckFlags,
@@ -31,8 +24,7 @@ pub(crate) fn check_verb(
     }
 }
 
-/// The `run` arm: a registry ref resolves to its verified cache file,
-/// then the run proceeds exactly as if given that path.
+/// `run` arm: resolve `registry:` then keep cache provenance.
 pub(crate) fn registry_then_run(
     mut args: RunArgs,
     color: ColorWhenArg,
@@ -40,8 +32,7 @@ pub(crate) fn registry_then_run(
     plain: bool,
     ascii: bool,
 ) -> u8 {
-    // The dispatcher's lazy resolver guaranteed the target before this
-    // seam — an absent file here is a wiring bug, surfaced honestly.
+    // Absent file here is a dispatcher wiring bug.
     let Some(file) = args.file.as_deref() else {
         return emit(&crate::VerbOutput {
             text: "nika run: no workflow target resolved (internal wiring)".to_owned(),
@@ -59,10 +50,7 @@ pub(crate) fn registry_then_run(
     }
 }
 
-/// Swap every registry ref among the file args for its verified cache
-/// path — non-refs pass untouched. `--fix` refuses refs BEFORE any
-/// network: a digest-pinned artifact stays read-only (rewriting the
-/// cache would poison its record).
+/// Resolve `registry:` args. `--fix` refuses refs before any network.
 fn resolve_registry_args(files: &[String], fix: bool) -> Result<Vec<CheckTarget>, VerbOutput> {
     if fix && files.iter().any(|f| registry::is_registry_ref(f)) {
         return Err(VerbOutput {
